fix: assumed "master" as default branch (#370)
* Disable nlreturn linter

`nlreturn` checks for a new line before return and branch statements to increase code clarity. This project does not adhere to this convention and fails `make lint` without this change.

* Change test to specify master branch

fribmendes/geometry changed its default branch to "main", and it no longer has a submodule to test on that branch. The master branch is still there though and has the async submodule to test, so this commit specifies that branch.

* Fix master as assumed branch (#369)

Instead of assuming "master" is always the branch we want, these changes allow for taking whatever default branch the GitHub repo specifies. This supports initiatives like the move to "main" as the default branch name.

* Add mattmc3's dotfiles to list